api/requirements.txt
beautifulsoup4==4.9.3
blis==0.7.3
Bottleneck==1.3.2
catalogue==1.0.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.4
fastai==1.0.60
fastprogress==1.0.0
ffmpeg-python==0.2.0
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.9
future==0.18.2
google-cloud-storage==1.37.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
murmurhash==1.0.4
numexpr==2.7.1
numpy==1.19.4
nvidia-ml-py3==7.352.0
packaging==20.7
pandas==1.1.4
Pillow==8.0.1
plac==1.1.3
preshed==3.0.4
pynvx==1.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
scipy==1.5.4
six==1.15.0
soupsieve==2.0.1
spacy==2.3.4
srsly==1.0.4
thinc==7.4.3
tqdm==4.54.0
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.0
Werkzeug==1.0.1
zipp==3.4.0